pyproject.toml
# This file was configured by CPA. For additional details: https://github.com/ysawa0/create-python-app
[tool.poetry]
name = "narrative"
version = "0.0.1"
description = ""
authors = [
  "Shayaan Jagtap <shayaan.jagtap@gmail.com>"
]
license = ""

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = false
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
color = true

[tool.isort]
balanced_wrapping = true
include_trailing_comma = true
known_first_party = "silver"
known_third_party = [
#  "boto3",        # Common for AWS
#  "django",       # Common web framework, if used
#  "flask",        # Common web framework, if used
#  "jinja2",       # Common templating engine
#  "matplotlib",   # Common for plotting
#  "numpy",        # Common for numerical operations
#  "pandas",       # Common for data manipulation
#  "pendulum",     # Common for date time
#  "pytest",       # Common for testing
#  "requests",     # Common for HTTP requests
#  "sqlalchemy",   # Common ORM for databases
]
multi_line_output = 3
profile = "black"
line_length = 120

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
gunicorn = "^21.2.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
cohere = "^4.50"
supabase = "^2.4.3"
marvin = "^2.3.4"
ping3 = "^4.0.8"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=3.5.0"
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"

[tool.pytest.ini_options]
pythonpath = ["src"]
